# api/docs/swagger.yaml
definitions:
...
  models.ComplianceProofResponse:
    properties:
      bpan:
        type: string
      commitment:
        items:
          type: integer
        type: array
      note:
        type: string
      proof:
        items:
          type: integer
        type: array
      requirement:
        description: '"operational", "second_life", "recyclable"'
        type: string
      statement:
        description: Human-readable claim
        type: string
    type: object
...
  models.TransferInitiateRequest:
    properties:
      reason:
        type: string
      to_owner_id:
        type: string
      to_owner_role:
        type: string
    type: object
  models.TransferRejectRequest:
    properties:
      reason:
        type: string
    type: object
  models.TransitionStateRequest:
    properties:
      details:
        type: string
      new_state:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/batteries/{bpan}/verify/operational:
    post:
      description: Generates a ZK proof that battery meets operational standards (government/regulator
        only)
      parameters:
      - description: BPAN
        in: path
        name: bpan
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ComplianceProofResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Verify battery operational compliance (ZK proof)
      tags:
      - compliance
...
  /batteries/{bpan}/ownership/history:
    get:
      description: Returns the full chain of custody for a battery
      parameters:
      - description: Battery PAN
        in: path
        name: bpan
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get ownership history
      tags:
      - lifecycle
  /batteries/{bpan}/ownership/transfer:
    post:
      consumes:
      - application/json
      description: Starts a dual-party consent ownership transfer
      parameters:
      - description: Battery PAN
        in: path
        name: bpan
        required: true
        type: string
      - description: Transfer payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.TransferInitiateRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Initiate battery ownership transfer
      tags:
      - lifecycle
...
  /batteries/{bpan}/recycling:
    post:
      consumes:
      - application/json
      description: Transitions state to RECYCLED (recycler only)
      parameters:
      - description: Battery PAN
        in: path
        name: bpan
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Record battery recycling
      tags:
      - lifecycle
  /batteries/{bpan}/reuse:
    post:
      consumes:
      - application/json
      description: Transitions state to SECOND_LIFE (reuse operator only)
      parameters:
      - description: Battery PAN
        in: path
        name: bpan
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Certify battery for second-life reuse
      tags:
      - lifecycle
...
  /batteries/{bpan}/transition:
    post:
      consumes:
      - application/json
      description: Moves a battery through the FSM (e.g. OPERATIONAL → SECOND_LIFE)
      parameters:
      - description: Battery PAN
        in: path
        name: bpan
        required: true
        type: string
      - description: Transition payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.TransitionStateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Transition battery lifecycle state
      tags:
      - lifecycle
  /batteries/{bpan}/verify/operational:
    post:
      description: Generates a zero-knowledge proof that battery SoH meets threshold
      parameters:
      - description: Battery PAN
        in: path
        name: bpan
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Verify battery is operational (ZK proof)
      tags:
      - lifecycle
  /batteries/{bpan}/verify/signature:
    post:
      description: Checks cryptographic integrity of all battery data
      parameters:
      - description: Battery PAN
        in: path
        name: bpan
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Verify battery data signature
      tags:
      - lifecycle
...
  /ownership/transfer/{id}/confirm:
    post:
      description: Confirms a pending transfer (either party). Transfer completes
        when both confirm.
      parameters:
      - description: Transfer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Confirm an ownership transfer
      tags:
      - lifecycle
  /ownership/transfer/{id}/reject:
    post:
      consumes:
      - application/json
      description: Rejects a pending transfer (either party can reject)
      parameters:
      - description: Transfer ID
        in: path
        name: id
        required: true
        type: string
      - description: Rejection reason
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.TransferRejectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Reject an ownership transfer
      tags:
      - lifecycle
swagger: "2.0"